chore: 컴포넌트 정리 및 스토리북 업데이트
