+++ src/components/buttons/radio-buttons/RadioButtons.stories.tsx
@@ -1,0 +1,18 @@
+import { Meta, StoryObj } from '@storybook/react';
+import RadioButtons from './RadioButtons';
+
+const meta: Meta<typeof RadioButtons> = {
+  title: 'Components/buttons/RadioButtons',
+  component: RadioButtons,
+  tags: ['autodocs'],
+};
+
+export default meta;
+
+type Story = StoryObj<typeof RadioButtons>;
+
+export const Primary: Story = {
+  args: {
+    names: ['여성', '남성'],
+  },
+};
