--- src/components/checkbox/CheckBox.stories.tsx
@@ -1,32 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react';
-import CheckBox from './CheckBox';
-
-const meta: Meta<typeof CheckBox> = {
-  title: 'Components/CheckBox',
-  component: CheckBox,
-  tags: ['autodocs'],
-};
-
-export default meta;
-
-type Story = StoryObj<typeof CheckBox>;
-
-export const Primary: Story = {
-  args: {
-    children: 'Primary Button',
-  },
-};
-
-export const Secondary: Story = {
-  args: {
-    children: 'Secondary Button',
-    variant: 'secondary',
-  },
-};
-
-export const Loading: Story = {
-  args: {
-    loading: true,
-    children: 'Loading',
-  },
-};
