--- src/components/chip/shuttle-status/ShuttleStatusChip.stories.tsx
@@ -1,16 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react';
-import ShuttleStatusChip from './ShuttleStatusChip';
-
-const meta: Meta<typeof ShuttleStatusChip> = {
-  title: 'Components/ShuttleStatusChip',
-  component: ShuttleStatusChip,
-  tags: ['autodocs'],
-};
-
-export default meta;
-
-type Story = StoryObj<typeof ShuttleStatusChip>;
-
-export const Primary: Story = {
-  args: {},
-};
