+++ src/components/inputs/text-input/TextInput.stories.tsx
@@ -1,0 +1,18 @@
+import { Meta, StoryObj } from '@storybook/react';
+import TextInput from './TextInput';
+
+const meta: Meta<typeof TextInput> = {
+  title: 'Components/inputs/TextInput',
+  component: TextInput,
+  tags: ['autodocs'],
+};
+
+export default meta;
+
+type Story = StoryObj<typeof TextInput>;
+
+export const Primary: Story = {
+  args: {
+    children: 'Primary Button',
+  },
+};
