--- src/components/radio/Radio.stories.tsx
@@ -1,16 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react';
-import Radio from './Radio';
-
-const meta: Meta<typeof Radio> = {
-  title: 'Components/Radio',
-  component: Radio,
-  tags: ['autodocs'],
-};
-
-export default meta;
-
-type Story = StoryObj<typeof Radio>;
-
-export const Primary: Story = {
-  args: {},
-};
