--- src/components/search/search-bar/SearchBar.stories.tsx
@@ -1,16 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react';
-import SearchBar from './SearchBar';
-
-const meta: Meta<typeof SearchBar> = {
-  title: 'Components/SearchBar',
-  component: SearchBar,
-  tags: ['autodocs'],
-};
-
-export default meta;
-
-type Story = StoryObj<typeof SearchBar>;
-
-export const Primary: Story = {
-  args: {},
-};
