--- src/components/text/text-area/TextArea.stories.tsx
@@ -1,16 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react';
-import TextArea from './TextArea';
-
-const meta: Meta<typeof TextArea> = {
-  title: 'Components/TextArea',
-  component: TextArea,
-  tags: ['autodocs'],
-};
-
-export default meta;
-
-type Story = StoryObj<typeof TextArea>;
-
-export const Primary: Story = {
-  args: {},
-};
